+++ cognite/neat/_graph/extractors/_classic_cdf/_relationships.py
@@ -10,8 +10,8 @@ from cognite.client.data_classes import Relationship, RelationshipList
 from rdflib import Namespace, URIRef
 
 from cognite.neat._issues.warnings import NeatValueWarning
-from cognite.neat._shared import Triple
 from cognite.neat._utils.auxiliary import create_sha256_hash
+from cognite.neat.core._shared import Triple
 
 from ._base import DEFAULT_SKIP_METADATA_VALUES, ClassicCDFBaseExtractor, InstanceIdPrefix, T_CogniteResource
