+++ cognite/neat/_graph/extractors/_dms.py
@@ -13,8 +13,8 @@ from cognite.neat._client import NeatClient
 from cognite.neat._config import GLOBAL_CONFIG
 from cognite.neat._constants import DEFAULT_SPACE_URI, is_readonly_property
 from cognite.neat._issues.errors import ResourceRetrievalError
-from cognite.neat._shared import Triple
 from cognite.neat._utils.collection_ import iterate_progress_bar
+from cognite.neat.core._shared import Triple
 
 from ._base import BaseExtractor
 from ._dict import DEFAULT_EMPTY_VALUES, DMSPropertyExtractor
