+++ cognite/neat/_graph/extractors/_mock_graph_generator.py
@@ -18,8 +18,8 @@ from cognite.neat._rules.models.data_types import DataType
 from cognite.neat._rules.models.entities import ClassEntity
 from cognite.neat._rules.models.information import InformationProperty
 from cognite.neat._rules.transformers import SubsetInformationRules
-from cognite.neat._shared import Triple
 from cognite.neat._utils.rdf_ import remove_namespace_from_uri
+from cognite.neat.core._shared import Triple
 
 from ._base import BaseExtractor
