+++ cognite/neat/_graph/queries/_select.py
@@ -8,8 +8,8 @@ from rdflib import Literal as RdfLiteral
 from rdflib.query import ResultRow
 
 from cognite.neat._constants import NEAT
-from cognite.neat._shared import InstanceType
 from cognite.neat._utils.rdf_ import remove_namespace_from_uri
+from cognite.neat.core._shared import InstanceType
 
 from ._base import BaseQuery
