+++ cognite/neat/core/_shared.py
