+++ tests/tests_unit/test_graph/test_transformers/test_attach_property_from_target2source.py
@@ -6,8 +6,8 @@ from rdflib import RDF, Literal, Namespace, URIRef
 
 from cognite.neat._constants import get_default_prefixes_and_namespaces
 from cognite.neat._graph.transformers._prune_graph import AttachPropertyFromTargetToSource
-from cognite.neat._shared import Triple
 from cognite.neat._store import NeatGraphStore
+from cognite.neat.core._shared import Triple
 
 
 def generate_test_parameters_delete_target_node() -> Iterable[ParameterSet]:
