+++ tests/tests_unit/test_graph/test_transformers/test_prune_graph.py
@@ -6,8 +6,8 @@ from rdflib import RDF
 
 from cognite.neat._constants import get_default_prefixes_and_namespaces
 from cognite.neat._graph.transformers import PruneDeadEndEdges
-from cognite.neat._shared import Triple
 from cognite.neat._store import NeatGraphStore
+from cognite.neat.core._shared import Triple
 
 
 def generate_test_parameters_unknown_types() -> Iterable[ParameterSet]:
